upgrade for rn 0.64 upgrade
after https://github.com/facebook/react-native/commit/be53aae324546ca038e2b58184ecc7066cfb598e,
`prepareHermes` is a Copy task. we should revert back as generic task to
support both copy and exec.
