+++ tools/src/versioning/android/versionReactNative.ts
@@ -85,3 +85,104 @@ async function runReactNativeCodegenAndroidAsync(
     'com.facebook.fbreact.specs',
   ]);
 }
+
+export async function renameHermesEngine(versionedReactAndroidPath: string, version: string) {
+  const abiVersion = version.replace(/\./g, '_');
+  const abiName = `abi${abiVersion}`;
+  const prebuiltHermesMkPath = path.join(
+    versionedReactAndroidPath,
+    'src',
+    'main',
+    'jni',
+    'first-party',
+    'hermes',
+    'Android.mk'
+  );
+  const versionedHermesLibName = `libhermes_${abiName}.so`;
+  await transformFileAsync(prebuiltHermesMkPath, [
+    {
+      find: /^(LOCAL_SRC_FILES\s+:=\s+jni\/\$\(TARGET_ARCH_ABI\))\/libhermes.so$/gm,
+      replaceWith: `$1/${versionedHermesLibName}`,
+    },
+  ]);
+
+  const buildGradlePath = path.join(versionedReactAndroidPath, 'build.gradle');
+  // patch prepareHermes task to rename copied library and update soname
+  // the diff is something like that:
+  //
+  // ```diff
+  // --- android/versioned-react-native/ReactAndroid/build.gradle.orig       2021-08-14 00:40:18.000000000 +0800
+  // +++ android/versioned-react-native/ReactAndroid/build.gradle    2021-08-14 00:40:58.000000000 +0800
+  // @@ -114,7 +114,7 @@
+  //      into("$thirdPartyNdkDir/folly")
+  //  }
+  //
+  // -task prepareHermes(dependsOn: createNativeDepsDirectories, type: Copy) {
+  // +task prepareHermes(dependsOn: createNativeDepsDirectories) {
+  //      def hermesPackagePath = findNodeModulePath(projectDir, "hermes-engine")
+  //      if (!hermesPackagePath) {
+  //          throw new GradleScriptException("Could not find the hermes-engine npm package", null)
+  // @@ -126,12 +126,29 @@
+  //      }
+  //
+  //      def soFiles = zipTree(hermesAAR).matching({ it.include "**/*.so" })
+  // -
+  // +    copy {
+  // +
+  //      from soFiles
+  //      from "src/main/jni/first-party/hermes/Android.mk"
+  //      into "$thirdPartyNdkDir/hermes"
+  // +
+  // +        rename '(.+).so', '$1_abi43_0_0.so'
+  // +    }
+  // +    exec {
+  // +        commandLine("patchelf", "--set-soname", "libhermes_abi43_0_0.so", "$thirdPartyNdkDir/hermes/jni/arm64-v8a/libhermes_abi43_0_0.so")
+  // +    }
+  // +    exec {
+  // +        commandLine("patchelf", "--set-soname", "libhermes_abi43_0_0.so", "$thirdPartyNdkDir/hermes/jni/armeabi-v7a/libhermes_abi43_0_0.so")
+  // +    }
+  // +    exec {
+  // +        commandLine("patchelf", "--set-soname", "libhermes_abi43_0_0.so", "$thirdPartyNdkDir/hermes/jni/x86/libhermes_abi43_0_0.so")
+  // +    }
+  // +    exec {
+  // +        commandLine("patchelf", "--set-soname", "libhermes_abi43_0_0.so", "$thirdPartyNdkDir/hermes/jni/x86_64/libhermes_abi43_0_0.so")
+  // +    }
+  //  }
+  //
+  // +
+  //  task downloadGlog(dependsOn: createNativeDepsDirectories, type: Download) {
+  //      src("https://github.com/google/glog/archive/v${GLOG_VERSION}.tar.gz")
+  //      onlyIfNewer(true)
+  // ```
+  await transformFileAsync(buildGradlePath, [
+    {
+      // reset `prepareHermes` task from Copy type to generic type then we can do both copy and exec.
+      find: /^(task prepareHermes\(dependsOn: .+), type: Copy(\).+$)/m,
+      replaceWith: '$1$2',
+    },
+    {
+      // wrap copy task and append exec tasks
+      find: /(^\s*def soFiles = zipTree\(hermesAAR\).+)\n([\s\S]+?)^\}/gm,
+      replaceWith: `\
+$1
+    copy {
+        $2
+        rename '(.+).so', '$$1_abi${abiVersion}.so'
+    }
+    exec {
+        commandLine("patchelf", "--set-soname", "${versionedHermesLibName}", "$thirdPartyNdkDir/hermes/jni/arm64-v8a/${versionedHermesLibName}")
+    }
+    exec {
+        commandLine("patchelf", "--set-soname", "${versionedHermesLibName}", "$thirdPartyNdkDir/hermes/jni/armeabi-v7a/${versionedHermesLibName}")
+    }
+    exec {
+        commandLine("patchelf", "--set-soname", "${versionedHermesLibName}", "$thirdPartyNdkDir/hermes/jni/x86/${versionedHermesLibName}")
+    }
+    exec {
+        commandLine("patchelf", "--set-soname", "${versionedHermesLibName}", "$thirdPartyNdkDir/hermes/jni/x86_64/${versionedHermesLibName}")
+    }
+}
+`,
+    },
+  ]);
+}
